+++ index.html
@@ -67,6 +67,6 @@
     </div>
 
     <script src="https://js.puter.com/v2/"></script>
-    <script src="main.js"></script>
+    <script type="module" src="main.js"></script>
 </body>
 </html>
